backAnalyst/requirements.txt
Django
djangorestframework
mysqlclient
python-decouple
paho-mqtt